+++ src/test/java/lotto/domain/MoneyTest.java
@@ -1,7 +1,8 @@
-package lotto.domain.manager;
+package lotto.domain;
 
 import static org.assertj.core.api.Assertions.*;
 
+import lotto.domain.Money;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
