+++ src/test/java/lotto/domain/TicketTest.java
@@ -1,7 +1,9 @@
-package lotto.domain.manager;
+package lotto.domain;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
+import lotto.domain.Money;
+import lotto.domain.Ticket;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
